pyproject.toml
[tool.poetry]
name = "trip-planner"
version="0.1.0"
description="Test Trip Planner"
authors=["Ben"]

[tool.poetry.dependencies]
python=">=3.10.0,<3.13"
crewai=">=0.1.24"
unstructured="==0.10.25"
pyowm='3.3.0'
python-dotenv="1.0.0"
langchain-openai = "^0.3.6"
langchain = "^0.3.19"
crawl4ai = "^0.4.248"

[tool.pyright]
useLibraryCodeForTypes=true
exclude=[".cache"]

[build-system]
requires=["poetry-code>=1.0.0"]
build-backend="poetry.core.masonry.api"

[tool.poetry.scripts]
crawl4ai_setup="crawl4ai-setup"